style: adjusted syle code
